requirements.txt
Django==1.7.6
django-filter==0.9.2
django-tastypie==0.12.1
djangorestframework==3.1.0
Markdown==2.6.1
psycopg2==2.6
python-dateutil==2.4.1
python-mimeparse==0.1.4
six==1.9.0
